+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Populate profile data automatically from ROKWIRE 2.0 Profile BB [#185](https://github.com/rokwire/core-building-block/issues/185)
+- Add phone authentication support [#24](https://github.com/rokwire/core-building-block/issues/24)
+- Handle anonymous ID conversion [#204](https://github.com/rokwire/core-building-block/issues/204)
 - Create a Security.md [#193] (https://github.com/rokwire/core-building-block/issues/193)
 - Set up authorization system [#45](https://github.com/rokwire/core-building-block/issues/45)
 - Add permissions to tokens [#189](https://github.com/rokwire/core-building-block/issues/189)
@@ -70,6 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixing the Log and the Changelog for issues #35 and #36 [#54](https://github.com/rokwire/core-building-block/issues/54)
 
 ### Changed
+- Login API issues [#182] (https://github.com/rokwire/core-building-block/issues/182)
 - Move temporary claims to auth library [#183](https://github.com/rokwire/core-building-block/issues/183)
 - Prepare the service to be deployed into Rokwire infrastructure [#176](https://github.com/rokwire/core-building-block/issues/176)
 - Users authentication polish [#155](https://github.com/rokwire/core-building-block/issues/155)